vendors/requirements.txt
Django
djangorestframework
psycopg2-binary
gunicorn
requests
django-rest-swagger
